djangorest/requirements.txt
dj-database-url==0.4.2
Django==1.11.7
djangorestframework==3.7.3
psycopg2==2.7.3.2
python-decouple==3.1
pytz==2017.3
whitenoise==3.3.1
